move logic from FileBackupConfig to base BackupConfig
- config validation
- get method
