+++ ebs_snapshots/backup_config.py
@@ -1,4 +1,20 @@
+import jsonschema
+import kayvee
+
+schema = {
+    "type": "object",
+    "name": "Volume backup schedule",
+    "properties": {
+            "max_snapshots": {"type": "number"},
+            "interval": {"type": "string"},
+            "name": {"type": "string"},
+    },
+    "additionalProperties": False,
+}
+
+
 class BackupConfig:
+
     """
     Interface shared by backup configs  
 
@@ -9,9 +25,26 @@ class BackupConfig:
         "vol-1234567" : {
             "interval": "daily",
             "max_snapshots": 7,
+            "name": "events-db",
             ...
         },
         ...
     }
     """
-    pass
+    @classmethod
+    def _validate_config(cls, new_config):
+        """ Raises exception if config loaded from file doesn't match expected schema """
+        assert type(new_config) is dict
+        for key, val in new_config.iteritems():
+            jsonschema.validate(val, schema)
+
+    def get(self):
+        """ Get a dict of config items """
+        try:
+            new_config = self.refresh()
+            self._validate_config(new_config)
+            self.config = new_config
+        except Exception as e:
+            print kayvee.formatLog("ebs-snapshots", "warning", "unable to load backup config", {"path": self.path, "error": str(e)})
+
+        return self.config
